Donution/pyproject.toml
[project]
name = "Donution"
version = "0.1.0"
description = "A web service for collecting donations for donuts and more."
authors = [
    {name = "Supersup66",email = "Alexey220V2014@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==5.2.1)",
    "djangorestframework (==3.16.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (==5.3.1)",
    "djoser (>=2.3.1,<3.0.0) ; python_version >= \"3.11\" and python_version < \"4.0\"",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pytest-lazy-fixtures (==1.3.4)",
    "faker (>=37.8.0,<38.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "gevent (>=25.9.1,<26.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "isort (>=6.1.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-django = "^4.11.1"
ruff = "^0.12.1"
psycopg2 = "^2.9.10"
pytest-cov = "^6.2.1"
responses = "^0.25.7"
